pyproject.toml
[project]
name = "law-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-anthropic>=1.3.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
]
